+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+- Option to disallow teleports to different worlds
+- Variable containing the distance between a player and waypoint for the cost formula (`distance`) 
+
 ## 4.5.3
 
 ### Added